MVP done
The main page now asks for a sentence.
When the sentence is given, pictures are automatically retrieved, of which 1 can be selected.
Subsequently, every word of the sentence can be clicked individually.
For each of the clicked words, a 'card' is added.
For cards that are verbs, the infinitive form is automatically added as a 'clue'.